refactor: 인터페이스들을 interfaces 패키지로 이동
