+++ src/test/java/attendance/domain/AttendanceBookTest.java
@@ -9,6 +9,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import attendance.exception.AttendanceArgumentException;
+import attendance.interfaces.SystemDateTime;
 
 class AttendanceBookTest {
     private final SystemDateTime systemDateTime = new AttendanceDateTime();
